+++ examples/tile-server/src/main.rs
@@ -1,3 +1,7 @@
+//! tile rendering server example
+//!
+//! It serves rendered tiles from a MapLibre style file via our pooling.
+
 use axum::{
     extract::Path,
     http::{header, StatusCode},
@@ -33,13 +37,13 @@ async fn rendered_style_tile(
         .as_image()
         .write_to(
             &mut std::io::Cursor::new(&mut png_bytes),
-            image::ImageFormat::Png,
+            image::ImageFormat::WebP,
         )
         .map_err(|e| dbg!(e))
         .map_err(|_| StatusCode::INTERNAL_SERVER_ERROR)?;
     let body = axum::body::Body::from(png_bytes);
     Ok(Response::builder()
-        .header(header::CONTENT_TYPE, "image/png")
+        .header(header::CONTENT_TYPE, "image/webp")
         .header(header::CACHE_CONTROL, "max-age=3600")
         .body(body)
         .unwrap())
@@ -54,8 +58,6 @@ async fn main() {
     let addr = "127.0.0.1:3000";
     println!("Server running on http://{addr}");
     let listener = tokio::net::TcpListener::bind(addr).await.unwrap();
-    let app = Router::new()
-        .route("/", get(index))
-        .route("/{z}/{x}/{y}", get(rendered_style_tile));
+    let app = Router::new().route("/", get(index)).route("/{z}/{x}/{y}", get(rendered_style_tile));
     axum::serve(listener, app).await.unwrap();
 }